Support for blaze command line flags for propeller_optimize feature.
propeller_optimize was added to blaze in https://github.com/bazelbuild/bazel/commit/2c7794b5053d849a45e4c73bc93e3004ce8e9e86. This change adds two
options to blaze to allow specifying the propeller optimize profile files as
part of the blaze command line.  The two options are:

1. --propeller_optimize_absolute_cc_profile
2. --propeller_optimize_absolute_ld_profile

The reason for this change is to allow BCID L3 builds from clients where
the BUILD file cannot be changed to include the generated profiles from a
different snapshot.

The behavior of these flags is the same as specifying the propeller_optimize
rule in the BUILD file.

This is a temporary change which must be refactored or deleted when
propeller_optimize is moved to Starlark.

PiperOrigin-RevId: 344838474