Don't return null for missing values if we have a pending exception to throw in package loading: it may be our last chance if we're in error bubbling.
Also don't unnecessarily wrap an IOException encountered during globbing: it leads to a less friendly error message since the exception message has "java.lang.IOException" in it. Similarly, pass the BuildFileNotFoundException message explicitly. That was exposed by this change because when we eagerly invalidate the package cache, that gives Skyframe globbing a chance to run and throw the slightly different exception.

PiperOrigin-RevId: 344901716