Allow Starlark transitions on --incompatible_enable_cc_toolchain_resolution
Add an exception to the transition function to allow --incompatible_enable_cc_toolchain_resolution to be an input and output to a transition.

Note that it is planned to convert this workaround into a commandline-specifiable allowlist if other arguments also need to excepted (e.g. incompatible_use_toolchain_resolution_for_java_rules). Some slight trickery will be required to convert this workaround to instead use that flag in the future, however.

RELNOTES: None
PiperOrigin-RevId: 345460229