Cleanup ConfiguredRuleClassProvider initialization a bit
We're not very disciplined about how we customize the ConfiguredRuleClassProvider (CRCP). There's:
- a setup() method that runs a bunch of mutations on a given builder,
- a create() static method that also provides the builder, and
- entry points in BlazeModule and tests that may use either method, combined with ad hoc additional initialization.

This CL factors some redundant initialization into the setup() method, which *should* be the canonical source of most business logic about how a CRCP is customized. However, additional logic may still be broken out into BlazeModules for pluggability, or to exclude the logic from our Java unit tests.

This CL also makes it easier to setup BuildViewTestCases that use a minimalistic CRCP. This is useful in BuiltinsInjectionTest, which tests basic functionality that 1) does not concern specific rules included in the default production CRCP, and 2) might be broken in the future by dependencies of those rules on builtins injection. Also includes a drive-by cleanup to split a bool arg into a separate method.

Work toward #11437.

PiperOrigin-RevId: 345513836