Automated rollback of commit 12b655b582c0d4eee5d7f62da62957336557cd96.
*** Reason for rollback ***

This will break anything that uses an --experimental_extra_action to analyze extra information about C++ compiles (e.g. IDE tooling). See CppCompileAction.getExtraActioninfo().

*** Original change description ***

Don't add discovered headers to an action's inputs. We already have a scheduling
dependency through a middleman and are still returning them from discoverInputs()
to properly declare Skyframe deps for invalidation. Without this change, the
current attempt to clearAdditionalInputs() is not helpful as a reference to the
created list of additional inputs is retained.

PiperOrigin-RevId: 345651734