Make it clearer that Args.add_all/add_joined process values lazily
The current wording says:

"For depsets, the items are evaluated lazily during the execution phase"

Which is true, the flattening of the depset with (effectively) depset.to_list() happens along with all the other processing at the start of the execution phase. But this makes it insufficiently clear that all the rest of the processing happens lazily for _both_ depsets and lists.

RELNOTES: None.
PiperOrigin-RevId: 346105700