Make `DottedVersion.Option` implement `StarlarkValue`.
Before this change, reading a value like `settings["//command_line_option:ios_minimum_os"]` from a Starlark transition threw a Java exception because the value is stored as an instance of `DottedVersion.Option`, not `DottedVersion`, and the former did not implement `StarlarkValue`.

Since the returned value is still the nested option type distinct from `DottedVersion`, the Starlark caller still needs to convert it to a string before use; this will return the original version number as written, without any canonicalization. The `apple_common.dotted_version` function can be used to turn that string into a `DottedVersion` that can be compared. For example:

```
def _some_transition_impl(settings, attr):
  version = apple_common.dotted_version(
    str(settings["//command_line_option:ios_minimum_os"]))
  if version >= apple_common.dotted_version("11.0"):
    # do something if iOS 11 or higher
```

PiperOrigin-RevId: 346126610