Stop hardcoding `-Xverify:none` in `SpawnAction`
The motivation was to improve startup performance of host tools that we trust.
The hard-coded flags for JavaBuilder and turbine were made obsolete by
`java_toolchain.jvm_opts`, which allows them to be configured in the toolchain
instead.

The option has been deprecated in Java 13 and will be removed in a future
release.

Fixes https://github.com/bazelbuild/bazel/issues/11381

PiperOrigin-RevId: 346160835