Make heuristic label expansion consistent regardless of the value of `--experimental_enable_aggregating_middleman`.
The test case added shows examples where a `filegroup` is passed to `tools`. Without middlemen, if it contains a single file, it is already substituted to that file's path. This change makes the substitution work with middlemen as well.

Additionally, when running without middlemen, if the `filegroup` contains zero or multiple files, we currently fail during heuristic expansion (analysis phase). This change ensures that an identical failure occurs with middlemen enabled. The action would presumably fail during execution anyway, but good to be consistent.

My benchmarking showed that the additional `NestedSet#toList` call is negligible on performance.

PiperOrigin-RevId: 346356033