Adds an api-android target to the sharding API.
Because Robolectric will run on desktops and emulators in the future, we need to add constraints = ["android"] to all targets that are Android-only.

I've added another target to the sharding API which satisfies this requirement, so that Dialer can use this target instead of the current one.

I had issues with getting the Bazel presubmit to recognize the `constraints` attribute on my target. Here is the YAQS answer that got me through this problem: []

PiperOrigin-RevId: 346375751