Convert BuildType.SelectorList back to SelectorList in native.existing_rules
This allows a round-trip like:

```
some_rule(
   ...
   deps = native.existing_rule(...)["deps"],
   ...
)
```

(native.existing_rule / native.existing_rules is disfavored, it's brittle and order-dependent. Prefer using a Starlark macro that creates the original target and related targets at the same time.)

When traversing the struct recursively, starlarkifyValue converts List into Tuple (though maybe StarlarkList would be more idiomatic) and Label into the corresponding str. So the select may not be exactly as input, but the meaning is preserved.

The logic for StarlarkValue is moved down because SelectorList is a StarlarkValue, so the more specific case needs to be handled first.

While this change is expected to be safe because rule instantiation cannot currently use these values as input, it's not strictly backward-compatible because it changes the type and `repr` of these values. For example, it's possible that rules could do something like:

```
if not 'SelectorList' in repr(native.existing_rule(...)[...]):
  # Something broken that wasn't previously executed
```

RELNOTES: native.existing_rule now returns select values in a form that is accepted by rule instantiation. This is a breaking API change, though the fallout is expected to be small.
PiperOrigin-RevId: 346555327