Starlark: Sequence.len protocol and fix range is not iterable
`range` need to be reimplemented using `StarlarkInt`.  For starters,
we can change the `Sequence` protocol to work with `StarlarkInt`,
not `int`.

This PR adds two operations to `Sequence` interface:
* `int lenInt() throws EvalException`, get size as `int`, throw on overflow
* `StarlarInt len()`, get size as `StarlarkInt`, does not throw

The first operation is used in sequence unpack, e. g. `[a, b] = l`.
The second operation is used in `len(x)` expression.

Before this PR:

```
>> range(-1 << 31, (1 << 31) - 1)
range(-2147483648, 2147483647)
>> len(range(-1 << 31, (1 << 31) - 1))
Traceback (most recent call last):
	File "<stdin>", line 1, column 4, in <toplevel>
Error in len: range is not iterable
```

"range is not iterable" error is reported because `RangeList.size`
returned negative integer.

Now:

```
>> range(-1 << 31, (1 << 31) - 1)
range(-2147483648, 2147483647)
>> len(range(-1 << 31, (1 << 31) - 1))
4294967295
```

We could fix range to explicitly handle size overflow in constructor,
but that's not the point, the point is future extension of range
to work with unbounded integers.