BEGIN_OSS starlark: make load statements bind file-locally
With this change, load statements create file-local bindings,
not global ones; a function that references a load binding
is a closure. There is no longer any need to maintain a
separate hash table in Module.getExportedGlobals that records
the "exported" globals.

The FileOptions.loadBindsGlobally option causes load statements
to create global bindings. It is used for the prelude
(which principally contains load statements), for the REPL,
where load statements in one chunk should be visible in the next,
and for BUILD files, which frequently, and for reasons unclear,
say "load(..., "x"); x=1".

This is a breaking Java API change for Copybara.
END_PUBLIC

PiperOrigin-RevId: 346753608