Automated rollback of commit 09f26fdadfad0e2e5fed4f377b564213d894bf49.
*** Reason for rollback ***

Only change to the include scanner, potentially the culprit of b/175294870

*** Original change description ***

Reduce needless contention in the include scanner's fileParseCache. The
computeIfAbsent() function of ConcurrentHashMap is supposed to have a fast
implementation and the map will then allow ~number of processors concurrent
writes. However, reading a file and extracting the include lines can be a costly
IO operation that does not fit this bill. Use futures instead.

RELNOTES: None.
PiperOrigin-RevId: 346800721