Propagate test envs to xml generation action
Previously, we hardcode the envs of the xml generation action, which
caused problem for process-wrapper because it's dynamically linked to
some system libraries and the required PATH or LD_LIBRARY_PATH are not
set.

This change propagate the envs we set for the actual test action to the
xml file generation action to make sure the env vars are correctly set and
can also be controlled by --action_env and --test_env.

Fixes https://github.com/bazelbuild/bazel/issues/4137
Fixes https://github.com/bazelbuild/bazel/issues/12579

Closes #12659.

PiperOrigin-RevId: 347596753