Make the navigation sidebar versioned according.
Fixes: #6937

This adds versioning support to the navigation sidebar in the default documentation page layout, conditionally loading a sidebar partial based on the current version's major version, or use the latest one in the case for rolling releases.

This gets rid of the many conditionals littered all over the navigation list items.