Document current providers interface for Target
Adjusts the summary to emphasize the struct-like nature of Target less, since that's only relevant for legacy providers. Add a bit of emphasis that this object represents a dependency.

Adjusts the documentation of Target.files to mention DefaultInfo and use the term "default outputs". That's more consistent with the terminology elsewhere in the documentation, and distinguishes that from other outputs files generated by the target (e.g. stuff that ends up in OutputGroupInfo or other providers but not in the DefaultInfo.files).

Renames "Extra providers" to "Providers". All providers are accessed in the new way. Among the other fields on Target, "label" and "aspect_ids" are not providers. "files" was a legacy provider, but now it's a field on DefaultInfo.

Removes class="anchor" from anchor elements. That seems redundant and I can't find it referenced in the site CSS.

Also fixes some formatting, fixes the link to ctx.aspect_ids (both the anchor and the field name include the "s"), and adds a link for struct.

RELNOTES: None.
PiperOrigin-RevId: 347849653