Add unregistered compile toolchains to system's JDK.
This was not foreseen by design document, but RBE and appengine want to use system JDK for compilation as well. Before hostjdk8 toolchain was used to achieve this. Just rewriting the toolchain into something that uses local_jdk version 8 does not achieve this objective.

So when a system JDK is imported, its version is detected and this also give us a chance to provide compilation toolchains for it. We keep them unregistered though, so the user can decide what to do.

Use case of using local JDK for compilation is achieved using flags:
```
--extra_toolchains=@local_jdk//:all  # add local JDK toolchains
--java_language_version=8            # Select to compile to Java language 8 (and will use local JDK for compilation)
```