"Inline" StarlarkSemantics into StarlarkBuiltinsValue
I.e., add the former as a field of the latter. You can still request the StarlarkSemantics in isolation, but if you request
StarlarkBuiltinsValue, you now get the semantics for free, without having to
incur an additional Skyframe edge.

Also added a constructor for making "empty" StarlarkBuiltinsValue instances
that wrap a semantics but do not contain any other information. This is handy
for providing a non-erroneous value to code paths where injection may or may
not be requested, without requiring the caller to separately and conditionally
request a semantics value.

BzlLoadFunction is refactored to thread along StarlarkBuiltinsValue in place
of StarlarkSemantics.

The memory savings of not taking an additional edge will be particularly
important when we make configured targets read the builtins value (for
exported_to_java).

Work toward #11437.

PiperOrigin-RevId: 347889368