Support running tests on the target platform
This change *includes* another change by @juliexxia that was previously
merged and rolled back:
https://github.com/bazelbuild/bazel/commit/c266ac966761c4b3d8a408a03e407505c93effdd

This adds an --use_target_platform_for_tests option that changes tests
to use the execution properties from the target platform rather than
the host platform. I believe that the code is currently incorrect -
if the host and target platforms differ, then tests are cross-compiled
for the target platform, but use the execution properties for the host
platform.

This matters for remote execution, where host and target platform may
be different CPU architectures and operating systems (e.g., compiling
on x64 Linux and running on ARM64 Mac). Currently, such tests will
typically fail to run if they contain architecture or OS-specific code.

Progress on #10799.

Change-Id: I774bd4442044d6725e78f496b9991368e73ffa00