Centralize control over the StarlarkThread in RuleContext
This lifts management of the StarlarkThread during rule analysis from
StarlarkRuleConfiguredTargetUtil to RuleContext. This allows the
StarlarkThread to be accessed in native rules that may wish to call
into Starlark code using the @_builtins mechanism.

The mutability is frozen at the end of the target's analysis, so multiple top-level calls to @_builtins-defined Starlark code within the same rule will share the same unfrozen mutability.

A follow-up will centralize management of the StarlarkRuleContext object as well.

Work toward #11437.

PiperOrigin-RevId: 348099327