More progress on #12081.
Create a capability to generate WORKSPACE files for tests at build time,
based on the repository definitions in distdir_deps.bzl.

This makes it possible for the versions of repositories used in tests
to align with the versions used a Bazel build time without upddating
files scattered throughout the source and test trees.

Demonstrated with rules_cc.