Generate rules/cpp/cc_configure.WORKSPACE from distdir_deps.bzl rather than keeping it in sync by hand
This makes it possible for the versions of repositories used in the distributed Bazel tests to align with the versions used at Bazel build time without updating files scattered throughout the source and test trees.

- Add //distdir_deps.bzl%gen_workspace_stanza rule to generate WORKSPACE rules for libraries in DISTDIR_DEPS.
- Apply the new capability to rules_cc.
- move dist_http_archive from distdir_deps.bzl to distdir.bzl so that all the repository time rules are distinct from the build time rules.

This finalizes the structural changes needed for issue #12081. The remaining part is applying the same pattern for each of the dependencies.

Closes #12743.

PiperOrigin-RevId: 350527359