Update Tar decompress function to handle symlinks after their targets
This PR updates the Tar archive decompress function to handle symbolic links entries only after their target is created.

If the symbolic link target does not exist on the time the link entry is being processed, we store the link entry so it will be processed once their target entry is extracted.

This is done to avoid a problem that happens on Windows where dangling junctions are created for file symlinks whose targets are part of the archive being decompressed but come after the link entry in the archive stream.

Fixes: #10399

Closes #12763.

PiperOrigin-RevId: 350936157