Add lexicographical sorting for graphless query unordered output guarded by a flag --incompatible_use_lexicographical_unordered_output.
The motivation behind this CL is to make consistent the output order between graphless query and graphless genquery, which outputs in unordered and lexicographical order respectively. The goal is to eventually switch to graphless query with lexicographical ordering as the default. This issue https://github.com/bazelbuild/bazel/issues/12757 keeps track of the flag on github.

PiperOrigin-RevId: 351107138