bazel analysis: preparatory cleanup to SRCTU error reporting
This change causes the error handling logic in
StarlarkRuleConfiguredTargetUtil (SRCTU) to avoid the
EvalException(Location) constructor. Instead, the auxiliary location,
if any, of provider instantiation is simply prepended to the error
message (see the infoError function) in anticipation of being printed
after a newline. For example:

   ERROR p/BUILD:1:1:\n
   foo.bzl:1:2: <provider-related message>

createTarget uses a distinct exception, BadImplementationFunction,
for errors that arise not from the Starlark implementation function
but from the post-processing of the providers.
A follow-up change will replace this exception by directly
reporting events to the handler, thus permitting both higher quality
structured errors capable of reporting more than one relevant location,
and multiple errors in a single pass. However, that change is too
tricky to accomplish in this CL.

Also:
- move "advertised provider" and "orphaned file" checks into createTarget.
- add comments to document this particularly painful code.
- delete EvalException.getDeprecatedLocation.
PiperOrigin-RevId: 351251301