Write common attributes in the specified alphabetical order
The expandCommonAttributes copied the map of attributes needlessly while modifying the RuleDocumentationAttribute objects themselves in-place. And it copied those attributes from a map which preserved insertion order to one that does not.

Currently, this results in the attributes being presented in alphabetical order, same as the lists in PredefinedAttributes.java.

RELNOTES: None
PiperOrigin-RevId: 351256718