skydoc: pay interest on fakebuildapi technical debt
commit 1f9b8ede2a59c3ef67d352f68278ca45fe6f115d caused symbols defined by load to no longer be recorded in
Module.globals, as the spec requires, but this broke the fragile
hacks used by the FakeStarlarkAttrModuleApi to guess the name of a
provider. This change causes the fakery to more faithfully simulate
Bazel's "export" mechanism (itself a hack) so that providers receive
a name when a top-level "name = provider()" statement is executed.

Also, all built-in providers that use the FakeProviderApi class (instead
of a specialized subclass) now explicitly specify their names at construction,
so that the heuristics get the correct name for nested fields such as cc_common.ToolchainInfo.

Nested fields that define specialized built-in providers still yield the
wrong value, but there are no tests that cover it and no complaints heard
so far, and this CL is penance enough already.

Fixes bazelbuild/stardoc#33

BUG=175703093
PiperOrigin-RevId: 351360957