bazel starlark transitions: grab build setting label from Starlark call stack in order to canonicalize.
There are 5 locations a user might write the label string of a build setting:
(1) on the command line
(2) in the transition definition inputs
(3) in the transition definition output
(4) in the transition impl function while reading the settings dict
(5) in the transition impl function return dict

We know that, depending on the situation, there are multiple ways to write "a build setting in the main repo of this project" - //myflag, @//myflag, @main_repo//my:flag.

In all 5 places listed above, bazel needs to recognize that the three different versions of //myflag above all map back to a single target. We need this for deduping purposes and consistency. 1 is taken care of in a grandparent CL to this CL. This Cl addresses locations 2-5.

To do this, whenever we are interpreting a user's input at one of the locations above, we covert that input string to the canonical string form of that build setting for our back end book keeping purposes. The logic to do this was modeled off of the `Label()` function logic. We also keep a map of canonicalized form to the user-given form in order to display messages (or key the dict in 4) to the user using the forms of the flags they used themselves.

work towards #11128

PiperOrigin-RevId: 351426279