Do not detect file names starting `..` as having up-level references.
`PathFragment::containsUplevelReferences` detects whether a path has an
up-level reference (`..`). The logic looks at the beginning of the normalized
path whether it starts with a `..`. This is incorrect for path segments which
have a `..` prefix, but are not `..`, e.g. `..file`. This causes bazel to crash
when we create a tree artifact with a file starting with a `..`.

Fix the detection to only recognize paths with `..` segments.

PiperOrigin-RevId: 351624019