Automated rollback of commit 52bf46b324a6bcc8f90f6b2ffaf6d877ff50ccf6.
*** Reason for rollback ***

b/177552620

*** Original change description ***

Reduce needless contention in the include scanner's fileParseCache. The
computeIfAbsent() function of ConcurrentHashMap is supposed to have a fast
implementation and the map will then allow ~number of processors concurrent
writes. However, reading a file and extracting the include lines can be a costly
IO operation that does not fit this bill. Use futures instead.

RELNOTES: None.
PiperOrigin-RevId: 351846931