Add the ability to implicitly enable split functions for FDO.
This change adds the ability to implicitly enable the "split_functions" feature
for instr. FDO targets when fdo_split_functions is enabled by default in the
future. This change adopts the same approach as ThinLTO to allow users to
opt-out via --features=-split_functions specified in the command line, package
or build rule.

Tested:
Added a unit test to check the various conditions after extending the mock
toolchain.

RELNOTES: None
PiperOrigin-RevId: 352027557