Gracefully print the error when --noenable_fdo_profile_absolute_path  is used.
blaze was crashing when --noenable_fdo_profile_absolute_path was used as it was
asserting on both profiles being null.  This change prevents the crash and
prints the problem as expected.

Further, when absolute path profiles are specified on the command line, do not
check for noenable_fdo_profile_absolute_path.  This is exactly how
cs_fdo_absolute_path also implements it and is no more unsafe than what already
exists.  propeller_optimize and cs_fdo_absolute_path are very similar in how
they consume profiles and what they do with it with regards to passing it down
to the compile action.

PiperOrigin-RevId: 352077037