Convert List to StarlarkList instead of Tuple for native.existing_rule
It's more common for those values to be passed in as lists, and that means the returned values are expressed in a more idiomatic way. Generally, inputs to rule attributes are expressed as lists instead of tuples, though either works.

However, this change is not backwards-compatible since tuple and list don't support precisely the same set of operations: Tuple can be used as a dict key if its contents are hashable, while list can't. Tuple can be concatenated with other tuples, but list can't.

RELNOTES: native.existing_rule now returns a mutable list, not a tuple, for a list-valued attributes. This is an incompatible API change.
PiperOrigin-RevId: 352551543